requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
sqlparse==0.4.3
tzdata==2022.7
## The following requirements were added by pip freeze:
amqp==5.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.5
attrs==22.2.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.0
billiard==4.2.1
bleach==6.0.0
celery==5.4.0
certifi==2025.6.15
cffi==1.15.1
charset-normalizer==3.4.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
django-cors-headers==4.0.0
env==0.1.0
executing==1.2.0
fastapi==0.85.2
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.8.0
fqdn==1.5.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.12.0
jsonpointer==2.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.5
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.27.3
kombu==5.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.26.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.20.4
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==2.5.0
uvicorn==0.19.0
vine==5.1.0
virtualenv==20.17.0
wcwidth==0.2.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7